@@ -5,7 +5,7 @@ import 'package:flutter_webapi_first_course/screens/home_screen/widgets/home_scr
 import '../../models/journal.dart';
 
 class HomeScreen extends StatefulWidget {
-  const HomeScreen({Key? key}) : super(key: key);
+  const HomeScreen({super.key});
 
   @override
   State<HomeScreen> createState() => _HomeScreenState();
