@@ -5,8 +5,7 @@ import 'package:flutter_webapi_first_course/models/journal.dart';
 class JournalCard extends StatelessWidget {
   final Journal? journal;
   final DateTime showedDate;
-  const JournalCard({Key? key, this.journal, required this.showedDate})
-      : super(key: key);
+  const JournalCard({super.key, this.journal, required this.showedDate});
 
   @override
   Widget build(BuildContext context) {
